backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.11
Pillow>=10.2.0
pytesseract==0.3.10
opencv-python-headless>=4.10.0
numpy>=1.26.0
scikit-learn>=1.5.0
pymupdf>=1.24.0
django-cors-headers==4.3.1
